@@ -18,7 +18,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -26,21 +26,21 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
-  font_awesome_flutter: ^8.11.0
-  google_fonts: ^1.1.1
-  flutter_svg: ^0.17.4
+  cupertino_icons: ^1.0.3
+  font_awesome_flutter: ^9.0.0
+  google_fonts: ^2.1.0
+  flutter_svg: ^0.22.0
   flutter_bloc: ^7.0.0
-  equatable: ^2.0.0
-  firebase_auth: ^0.18.4+1
-  google_sign_in: ^4.4.4
-  firebase_core: ^0.5.0
-  flutter_search_bar: ^2.1.0
-  carousel_slider: ^1.3.0
-  shared_preferences: ^0.5.12+4
-  http: ^0.12.2
-  material_floating_search_bar: ^0.3.3
-  url_launcher: ^6.0.3
+  equatable: ^2.0.2
+  firebase_auth: ^1.2.0
+  google_sign_in: ^5.0.3
+  firebase_core: ^1.2.0
+  flutter_search_bar: ^3.0.0-dev.1
+  carousel_slider: ^4.0.0-nullsafety.0
+  shared_preferences: ^2.0.5
+  http: ^0.13.3
+  material_floating_search_bar: ^0.3.4
+  url_launcher: ^6.0.4
   flutter_staggered_grid_view: ^0.4.0
 
 dev_dependencies:
