@@ -1,8 +1,8 @@
 import 'package:firebase_auth/firebase_auth.dart';
 
 class LoginResponse {
-  final User user;
+  final User? user;
   final String message;
 
-  LoginResponse({this.user, this.message});
+  LoginResponse({required this.user, required this.message});
 }
