@@ -1,8 +1,8 @@
 import 'package:firebase_auth/firebase_auth.dart';
 
 class RegisterResponse {
-  final User user;
+  final User? user;
   final String message;
 
-  RegisterResponse({this.user, this.message});
+  RegisterResponse({required this.user, required this.message});
 }
