@@ -11,7 +11,7 @@ class LoginAccount extends LoginEvent {
   final String email;
   final String password;
 
-  LoginAccount({@required this.email, @required this.password});
+  LoginAccount({required this.email, required this.password});
 
   @override
   List<Object> get props => [email, password];
@@ -21,7 +21,7 @@ class RegisterAccount extends LoginEvent {
   final String email;
   final String password;
 
-  RegisterAccount({@required this.email, @required this.password});
+  RegisterAccount({required this.email, required this.password});
 
   @override
   List<Object> get props => [email, password];
