@@ -16,7 +16,7 @@ class LoginSuccess extends LoginState {}
 class LoginFailure extends LoginState {
   final String message;
 
-  LoginFailure({@required this.message});
+  LoginFailure({required this.message});
 
   @override
   List<Object> get props => [message];
@@ -27,7 +27,7 @@ class RegisterLoading extends LoginState {}
 class RegisterSuccess extends LoginState {
   final String message;
 
-  RegisterSuccess({@required this.message});
+  RegisterSuccess({required this.message});
 
   @override
   List<Object> get props => [message];
@@ -36,7 +36,7 @@ class RegisterSuccess extends LoginState {
 class RegisterFailure extends LoginState {
   final String message;
 
-  RegisterFailure({@required this.message});
+  RegisterFailure({required this.message});
 
   @override
   List<Object> get props => [message];
