@@ -4,17 +4,21 @@ import 'package:pricer_project/models/register_response.dart';
 
 class UserRepository {
   final PricerApi pricerApi;
-  UserRepository({this.pricerApi});
+  UserRepository({required this.pricerApi});
 
   // user login
-  Future<LoginResponse> signInWithEmailAndPassword({String email, String password}) async {
-    LoginResponse loginResponse = await pricerApi.signInWithEmailAndPassword(email: email, password: password);
+  Future<LoginResponse> signInWithEmailAndPassword(
+      {required String email, required String password}) async {
+    LoginResponse loginResponse = await pricerApi.signInWithEmailAndPassword(
+        email: email, password: password);
     return loginResponse;
   }
 
   // user register
-  Future<RegisterResponse> registerWithEmailAndPassword({String email, String password}) async {
-    RegisterResponse registerResponse = await pricerApi.registerWithEmailAndPassword(email: email, password: password);
+  Future<RegisterResponse> registerWithEmailAndPassword(
+      {required String email, required String password}) async {
+    RegisterResponse registerResponse = await pricerApi
+        .registerWithEmailAndPassword(email: email, password: password);
     return registerResponse;
   }
 }
