@@ -5,7 +5,7 @@ class SimpleData {
   final Related related;
   final List<TokpedProduct> products;
 
-  SimpleData({this.related, this.products});
+  SimpleData({required this.related, required this.products});
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = Map<String, dynamic>();
@@ -17,7 +17,8 @@ class SimpleData {
   factory SimpleData.fromJson(Map<String, dynamic> json) {
     if (json['products'] != null) {
       var tagObjsJson = json['products'] as List;
-      List<TokpedProduct> _product = tagObjsJson.map((e) => TokpedProduct.fromJson(e)).toList();
+      List<TokpedProduct> _product =
+          tagObjsJson.map((e) => TokpedProduct.fromJson(e)).toList();
 
       return SimpleData(
         related: Related.fromJson(json['related']),
