@@ -6,24 +6,28 @@ import 'package:pricer_project/models/register_response.dart';
 class FireBaseAuthService {
   static FirebaseAuth _auth = FirebaseAuth.instance;
 
-  static Future<LoginResponse> signInWithEmailAndPassword({String email, String password}) async {
+  static Future<LoginResponse> signInWithEmailAndPassword(
+      {required String email, required String password}) async {
     await Firebase.initializeApp();
     try {
-      UserCredential userCredential = await _auth.signInWithEmailAndPassword(email: email, password: password);
-      return LoginResponse(user: userCredential.user);
+      UserCredential userCredential = await _auth.signInWithEmailAndPassword(
+          email: email, password: password);
+      return LoginResponse(user: userCredential.user, message: '');
     } catch (e) {
-      return LoginResponse(message: e.toString());
+      return LoginResponse(user: null, message: e.toString());
     }
   }
 
-  static Future<Object> registerWithEmailAndPassword({String email, String password}) async {
+  static Future<RegisterResponse> registerWithEmailAndPassword(
+      {required String email, required String password}) async {
     await Firebase.initializeApp();
     try {
-      UserCredential userCredential = await _auth.createUserWithEmailAndPassword(email: email, password: password);
-      return RegisterResponse(user: userCredential.user);
+      UserCredential userCredential = await _auth
+          .createUserWithEmailAndPassword(email: email, password: password);
+      return RegisterResponse(user: userCredential.user, message: '');
     } catch (e) {
       print(e);
-      return RegisterResponse(message: e.toString());
+      return RegisterResponse(user: null, message: e.toString());
     }
   }
 }
