@@ -2,8 +2,6 @@ import 'dart:async';
 
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
 import 'package:pricer_project/data/repositories/user_repositories.dart';
 import 'package:pricer_project/logic/auth/auth_bloc.dart';
 import 'package:pricer_project/models/login_response.dart';
@@ -16,7 +14,8 @@ class LoginBloc extends Bloc<LoginEvent, LoginState> {
   final UserRepository userRepository;
   final AuthBloc authBloc;
 
-  LoginBloc({this.userRepository, this.authBloc}) : super(LoginInitial());
+  LoginBloc({required this.userRepository, required this.authBloc})
+      : super(LoginInitial());
 
   @override
   Stream<LoginState> mapEventToState(
@@ -34,12 +33,16 @@ class LoginBloc extends Bloc<LoginEvent, LoginState> {
     yield LoginLoading();
 
     try {
-      LoginResponse loginResponse = await userRepository.signInWithEmailAndPassword(email: event.email, password: event.password);
+      LoginResponse loginResponse =
+          await userRepository.signInWithEmailAndPassword(
+              email: event.email, password: event.password);
       if (loginResponse.message.toLowerCase() == 'login success') {
         authBloc.add(UserLoggedIn(userEmail: event.email));
         yield LoginSuccess();
       } else {
-        yield LoginFailure(message: loginResponse.message.replaceAll(new RegExp(r'[\(\[].*?[\)\]]'), ''));
+        yield LoginFailure(
+            message: loginResponse.message
+                .replaceAll(new RegExp(r'[\(\[].*?[\)\]]'), ''));
       }
     } catch (e) {
       print(e.toString());
@@ -51,11 +54,15 @@ class LoginBloc extends Bloc<LoginEvent, LoginState> {
     yield LoginLoading();
 
     try {
-      RegisterResponse registerResponse = await userRepository.registerWithEmailAndPassword(email: event.email, password: event.password);
+      RegisterResponse registerResponse =
+          await userRepository.registerWithEmailAndPassword(
+              email: event.email, password: event.password);
       if (registerResponse.message.toLowerCase() == 'register success') {
         yield RegisterSuccess(message: registerResponse.message);
       } else {
-        yield RegisterFailure(message: registerResponse.message.replaceAll(new RegExp(r'[\(\[].*?[\)\]]'), ''));
+        yield RegisterFailure(
+            message: registerResponse.message
+                .replaceAll(new RegExp(r'[\(\[].*?[\)\]]'), ''));
       }
     } catch (e) {
       print(e.toString());
