@@ -11,7 +11,7 @@ class GetTokpedSuggestion extends SearchEvent {
   final String query;
   final String limit;
 
-  GetTokpedSuggestion({this.query, this.limit});
+  GetTokpedSuggestion({required this.query, required this.limit});
 
   @override
   List<Object> get props => [query, limit];
@@ -23,7 +23,12 @@ class GetTokpedProduct extends SearchEvent {
   final String limit;
   final bool fromLow;
 
-  GetTokpedProduct({this.query, this.limit, this.fromLow, this.parentCategory});
+  GetTokpedProduct({
+    required this.query,
+    required this.limit,
+    required this.fromLow,
+    required this.parentCategory,
+  });
 
   @override
   List<Object> get props => [query, limit];
