@@ -1,25 +1,26 @@
 import 'package:flutter/material.dart';
 
 class WidgetNotificationSnackbar {
-  final Color color;
-  final String message;
+  WidgetNotificationSnackbar();
 
-  WidgetNotificationSnackbar({this.color, this.message});
-
-  render({BuildContext context, Color color, String message}) {
+  render({
+    required BuildContext context,
+    required Color color,
+    required String message,
+  }) {
     // ignore: deprecated_member_use
     return Scaffold.of(context).showSnackBar(
       SnackBar(
         behavior: SnackBarBehavior.floating,
         margin: EdgeInsets.all(15),
         duration: Duration(milliseconds: 1200),
-        backgroundColor: color ?? Colors.green,
+        backgroundColor: color,
         shape: RoundedRectangleBorder(
           borderRadius: BorderRadius.all(
             Radius.circular(10),
           ),
         ),
-        content: Text(message ?? ''),
+        content: Text(message),
       ),
     );
   }
